separate tests for log_figure
